stop retrying if tasks aren't making any progress
coding style fixed

optimize UploadFailureItemTask and fix coding style

fix the race condition when re-initializing and optimize code.

retrieve encKey from ContentDb

make codes more succinct

+++ app/src/main/java/com/halloapp/media/ResumableUploader.java
@@ -53,7 +53,7 @@ public class ResumableUploader {
     }
 
     @WorkerThread
-    public static String sendPatchRequest(File file, int offset, @NonNull String url, @Nullable ResumableUploadListener listener) throws IOException {
+    public static String sendPatchRequest(File file, long offset, @NonNull String url, @Nullable ResumableUploadListener listener) throws IOException {
         final HttpURLConnection connection = (HttpURLConnection) new URL(url).openConnection();
 
         connection.setChunkedStreamingMode(BUFFER_SIZE);
